refactor: move ElevationGuard into shared route infrastructure
The previous commit documented the cross-module guard import as an accepted
deviation, which amounted to amending the rule in the same diff that broke it.
Doing the consolidation instead, scoped so it does not disturb the twenty-odd
existing `AuthGuard` consumers.

`AUTH_PAYLOAD_REQUEST_PROPERTY` moves to `src/routes/common/auth/`, where it is
a plain string constant with no imports of its own. `AuthGuard` keeps its static
alias, so every existing reference and the Fastify request augmentation are
untouched. `ElevationGuard` moves next to it and reads the shared constant, so
it now imports only `@/config`, `@/domain/common`, the auth module's `domain/`
and `@/routes/common` — the five spaces controllers and the route matrix reach
it through shared infrastructure rather than another module's `routes/` tree.

`docs/agents/module-structure.md` goes back to what it said before: the
deviation the anti-example names is not precedent, and this change no longer
needs it to be.

Consolidating `AuthGuard` and `@Auth()` the same way is still worth doing and
is still its own change — this one only stops adding to the pile.
