+++ docs/agents/module-structure.md
@@ -80,8 +80,6 @@ This applies in both directions: neither a module's `routes/` nor its `datasourc
 
 **Anti-example:** `src/modules/billing/routes/billing.controller.ts` imports `@/modules/spaces/routes/pipes/space-id.pipe` and `@/modules/auth/routes/decorators/auth.decorator`; `src/modules/notifications/domain/v2/entities/notification.entity.ts` imports from `@/modules/hooks/routes/entities/event-type.entity`. Neither is precedent — do not imitate.
 
-**Named exception — the auth module's route-layer primitives.** `AuthGuard`, `ElevationGuard` and the `@Auth()` decorator live in `src/modules/auth/routes/` and are imported from there by every module with an authenticated route (23 files). This is a deviation, and it is the one the anti-example above names for `auth.decorator` — but it is a *recorded* deviation rather than a fresh one, because the rule has no compliant target for a Nest guard: `domain/` holds interfaces and entities and must not know about `ExecutionContext`, and relocating them to `src/routes/common/` would invert the dependency, making shared infrastructure depend on a feature module's payload contract. The rule's stated rationale — keeping HTTP-layer concerns out of code that has no business depending on them — does not bite here either, since every consumer is itself a controller. Non-negotiable #9 requires `ElevationGuard` on sensitive Workspace routes, so it is imported the same way as `AuthGuard` beside it. Consolidating all three into shared infrastructure is the real fix and is its own change; until then, do not add *other* kinds of cross-module `routes/` import on the strength of this entry.
-
 ### Frozen legacy trees
 
 **Rule:** Do not add new files under `src/routes/`, except to the shared infrastructure in `src/routes/common/`. Do not add new feature entities to `src/domain/`. Do not add new API-client directories under central `src/datasources/` — a new client is a module-local `datasources/`. Central `src/datasources/` holds only cross-cutting infra (cache, db, network, jwt, kms, job-queue, storage, circuit-breaker) plus six legacy API clients that may still be imported but never joined by a new sibling: `billing-api`, `config-api`, `etherscan-api`, `fee-service-api`, `locking-api`, `push-notifications-api`.
